requirements.txt
matplotlib==3.5.3
numpy==1.23.5
pandas==1.5.3
pillow==9.4.0
streamlit==1.29.0
